// packages/extension/src/protocol/plugin.d.ts
export interface Request {
  initialize?: InitializeRequest;
  onchange?: OnChangeRequest;
  [k: string]: unknown;
}
export interface InitializeRequest {
  capabilities?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface OnChangeRequest {
  context?: Context;
  change?: Change;
  [k: string]: unknown;
}
export interface Context {
  /**
   * The path to the client's currently focused file.
   */
  filePath?: string;
  /**
   * The content of the focused file.
   */
  fileContent?: string;
  /**
   * The Sysl module encoded in the focused file.
   */
  module?: string;
  /**
   * The path to the Sysl root.
   */
  syslRoot?: string;
  /**
   * The specific position on which the client is focused.
   */
  focus?: {
    src?: string;
    app?: string[];
    type?: string;
    field?: string;
    anno?: string;
    tag?: string;
    endpoint?: string;
    stmt?: string;
    param?: string;
    [k: string]: unknown;
  };
  /**
   * The state of relevant settings in the client.
   */
  settings?: {
    [k: string]: unknown;
  };
  /**
   * The ID of the view where the change occurred.
   */
  viewId?: string;
  [k: string]: unknown;
}
/**
 * Represents a single change to a single file in the client that the plugin may react to.
 */
export interface Change {
  /**
   * The surface in the client where the change originated.
   */
  source?: "UNSPECIFIED_SOURCE" | "TEXT" | "DIAGRAM" | "CUSTOM";
  /**
   * The action that caused the change.
   */
  action?:
    | "UNSPECIFIED_ACTION"
    | "ADD"
    | "MODIFY"
    | "REMOVE"
    | "CREATE_FILE"
    | "SAVE_FILE"
    | "DELETE_FILE";
  /**
   * The path to the file that changed.
   */
  filePath?: string;
  detail?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
/**
 * General wrapper for all responses.
 */
export interface Response {
  error?: Error;
  initialize?: InitializeResponse;
  onchange?: OnChangeResponse;
  [k: string]: unknown;
}
export interface Error {
  code?: number;
  message?: string;
  data?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface InitializeResponse {
  capabilities?: ServerCapabilities;
  [k: string]: unknown;
}
export interface ServerCapabilities {
  diagrams?: DiagramCapabilities;
  [k: string]: unknown;
}
// ...
export interface OnChangeResponse {
  renderDiagram?: Diagram[];
  [k: string]: unknown;
}
export interface Diagram {
  type?: {
    id?: string;
    name?: string;
    description?: string;
    icon?: string;
    [k: string]: unknown;
  };
  content?: DiagramData;
  [k: string]: unknown;
}